+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+- now you can add screen and controller with command
+
 ## 1.0.0
 
-- Initial version, created by Stagehand
+- Project Created